requirements.txt
streamlit
python-dotenv
gspread
google-auth
pypdf
langchain-google-genai
langgraph
typing_extensions